refactored entire codebase
