+++ .eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "@typescript-eslint/no-explicit-any": "warn"
+  }
+}
