+++ src/types/channel.ts
@@ -1,0 +1,7 @@
+export interface Channel {
+  identifier: string;
+  name: string;
+  code: string;
+  status: string;
+  lastUpdatedOn?: string;
+}
